Initial Commit
For a group size n and multiplier t, this will return the cyclotomic
cosets (orbits) of the group and multiplier.  Returns a matrix whose
first entry is the identity element (0) and then each following row is a
particular orbit.  The first column is the complete set of
representatives.  Any 0s not in the first row are place holders.